Added naming functionality for the results in the result window.
--Name:
is the tag to name a select, and it will name the result in the order or appearence, ie, for this result the first tag, for the second result, the next tag, and so on.
This means that if you want all the results named is posible, just a the first "n" results named is also posible, but you can skip a name on a query or it will be named most likely wrong.